# book-network-ui/src/app/services/models.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export { type AuthenticationRequest } from './models/authentication-request';
export { type AuthenticationResponse } from './models/authentication-response';
export { type BookRequest } from './models/book-request';
export { type BookResponse } from './models/book-response';
export { type BorrowedBookResponse } from './models/borrowed-book-response';
export { type FeedbackRequest } from './models/feedback-request';
export { type FeedbackResponse } from './models/feedback-response';
export { type PageResponseBookResponse } from './models/page-response-book-response';
export { type PageResponseBorrowedBookResponse } from './models/page-response-borrowed-book-response';
export { type PageResponseFeedbackResponse } from './models/page-response-feedback-response';
export { type RegistrationRequest } from './models/registration-request';